WoT restructure A5: create rangeQuery/ (renderedRange + sticky mixins from viewport/ + table/mixin/)
