+++ handsontable/src/3rdparty/walkontable/src/rangeQuery/renderedRange.ts
