+++ handsontable/src/3rdparty/walkontable/src/table.ts
@@ -1,7 +1,7 @@
 import type { WalkontableInstance } from './types';
 import type { EngineContext } from './wire';
 import type Settings from './settings';
-import type { RowRangeQuery, ColumnRangeQuery } from './viewport/renderedRange';
+import type { RowRangeQuery, ColumnRangeQuery } from './rangeQuery/renderedRange';
 import { cellAccess, type CellAccess } from './cellLocator/cellAccess';
 import { domScaffold, type DomScaffold } from './table/domScaffold';
 import {
