+++ handsontable/src/3rdparty/walkontable/src/table/bottomInlineStartCorner.ts
@@ -1,7 +1,7 @@
 import type { TableDeps } from '../table';
 import Table from '../table';
-import stickyRowsBottom from './mixin/stickyRowsBottom';
-import stickyColumnsStart from './mixin/stickyColumnsStart';
+import stickyRowsBottom from '../rangeQuery/stickyRowsBottom';
+import stickyColumnsStart from '../rangeQuery/stickyColumnsStart';
 import { mixin } from '../../../../helpers/object';
 import { CLONE_BOTTOM_INLINE_START_CORNER } from '../overlay';
